+++ src/main/java/com/uniara/inventory/domain/parts/PartsRequest.java
@@ -4,5 +4,5 @@ import com.uniara.inventory.domain.dealer.Dealer;
 
 import java.math.BigDecimal;
 
-public record PartsRequest(String partNumber, Integer quantity, BigDecimal value) {
+public record PartsRequest(String partNumber, Integer quantity, Integer value) {
 }
