+++ src/main/java/com/uniara/inventory/domain/parts/PartsResponse.java
@@ -6,7 +6,7 @@ import com.uniara.inventory.domain.parts.Parts;
 
 import java.math.BigDecimal;
 
-public record PartsResponse(String partNumber, Integer quantity, BigDecimal value, Dealer dealer) {
+public record PartsResponse(String partNumber, Integer quantity, Integer value, Dealer dealer) {
         public PartsResponse(Parts parts){
             this(parts.getPartNumber(), parts.getQuantity(), parts.getValue(), parts.getDealer());
         }
